Move xr_zeros to odc.geo.xr.
